BaseWhiteNoise class
