+++ whitenoise/base.py
@@ -3,12 +3,10 @@ from __future__ import annotations
 import os
 import re
 import warnings
-from posixpath import normpath
 from wsgiref.headers import Headers
-from wsgiref.util import FileWrapper
 
 from .media_types import MediaTypes
-from .responders import IsDirectoryError, MissingFileError, Redirect, StaticFile
+from .responders import MissingFileError, Redirect, StaticFile
 from .string_utils import (
     decode_if_byte_string,
     decode_path_info,
@@ -16,7 +14,7 @@ from .string_utils import (
 )
 
 
-class WhiteNoise:
+class BaseWhiteNoise:
 
     # Ten years is what nginx sets a max age if you use 'expires max;'
     # so we'll follow its lead
@@ -74,28 +72,6 @@ class WhiteNoise:
             self.immutable_file_test = lambda path, url: bool(regex.search(url))
         if root is not None:
             self.add_files(root, prefix)
-
-    def __call__(self, environ, start_response):
-        path = decode_path_info(environ.get("PATH_INFO", ""))
-        if self.autorefresh:
-            static_file = self.find_file(path)
-        else:
-            static_file = self.files.get(path)
-        if static_file is None:
-            return self.application(environ, start_response)
-        else:
-            return self.serve(static_file, environ, start_response)
-
-    @staticmethod
-    def serve(static_file, environ, start_response):
-        response = static_file.get_response(environ["REQUEST_METHOD"], environ)
-        status_line = f"{response.status} {response.status.phrase}"
-        start_response(status_line, list(response.headers))
-        if response.file is not None:
-            file_wrapper = environ.get("wsgi.file_wrapper", FileWrapper)
-            return file_wrapper(response.file)
-        else:
-            return []
 
     def add_files(self, root, prefix=None):
         root = decode_if_byte_string(root, force_text=True)
